Use FNV HashMap over std::collections where applicable
Initially, I chose `BTreeMap` because many benchmarks indicate that it
tends to be much faster than `HashMap` in many cases due to the default
DDoS-resistant hasher actually being slower than computing and
traversing a B-tree. However, if we switch to a faster hashing algorithm
it should perform much better, albeit at the cost of the aforementioned
DDoS resistance.