Add App abstraction for managing init and main loop
This allows us to build generic UI applications using our widget cache
and manage the main loop.

Note that this main loop current relies on the monitor V-sync to keep
the framerate consistent. Ideally, we should try to fix our timestep in
some other way, e.g. using `std::thread::sleep()` or using an
accumulator. The former is the simplest approach to accomplish this, but
results may be not so consistent, as the resolution for most OS sleep
functions are in the order of 1-2 milliseconds, which may be too
granular for our needs. The latter approach is more accurate and more
efficient, but requires a little more work.

Since we still need to define our widgets and get something concrete
rendering to the screen, we will put this aside for now and focus on
fixing our timestep in a later commit.