Create and store base texture for each widget
Instead of having each `W: Widget` type manage its own base texture, as
originally intended, have the `Widgets<W>` cache store it instead. If
the `Widget::properties().bounds` field is updated, the texture is
automatically resized if necessary before `Widget::draw()` is called.

This change was made due to the `sdl2::render::Texture` having an
associated lifetime tied to the `sdl2::render::TextureCreator` that
created it. If this design change was not done, every `Widget`
implementation would be required to specify a `'tc` lifetime, which
would be terrible for API ergonomics.