Disable 'bundled' and 'static-link', add README.md
It turns out that the "static-link" feature flag for the `sdl2-sys`
crate doesn't link `SDL2_Image` nor `SDL2_TTF`) into it, and the
"bundled" feature flag only fetches the main `libsdl2` library at build
time and not `SDL2_Image`. Plus, it seems that "bundled" and
"static-link" are also broken in other ways, at the time of writing:

https://github.com/Rust-SDL2/rust-sdl2/issues/778
https://github.com/Rust-SDL2/rust-sdl2/issues/851
https://github.com/Rust-SDL2/rust-sdl2/issues/858
https://github.com/Rust-SDL2/rust-sdl2/issues/1017
https://github.com/Rust-SDL2/rust-sdl2/issues/1029
https://github.com/Rust-SDL2/rust-sdl2/issues/1034

For the sake of reliability, this commit disables both Cargo feature
flags described for the `sdl2` dependency (thereby requiring the user to
install external dependencies) and introduces a detailed setup guide in
the `README.md` for Windows, macOS, and Linux.