Change Widget::draw() to take &mut self
Ideally, `Widget::draw()` should be a pure function which draws the
current state of the widget without mutating its internal state.
However, there may be a few occasions where we need to violate this
principle, namely when we need to "pop in" a tile thumbnail image, once
it's done downloading. We must store the rendered texture inside the
widget, and that requires changing the method signature to take `&mut
self`.