Implement lazy fetching/loading of tile thumbnails
NOTE: This commit temporarily reverts back to using a local (patched)
copy of `home.json` in order to work around an apparent API bug
preventing some images from loading (I've reported upstream by email).

There seem to be a few typos in the values of the `masterId` and `url`
keys of this JSON object:

```
.data.StandardCollection.containers[0].set.items[1].image.tile."1.78".series.default
```

At the time of writing, the given image thumbnail URL is reported as:

https://prod-ripcut-delivery.disney-plus.net/v1/variant/disney/E7BFFF8CD4E7BA85BFB3439CDF90698213E8134E4CC729A9AFA17A2E1FC665D31/scale?format=jpeg&quality=90&scalingAlgorithm=lanczos3&width=500

This page returns a 404 because the master ID contains an unnecessary 1
digit at the end, thereby making it 65 characters long instead of 64
(the `masterId` field has an identical mistake). It seems that all other
images in `.data.StandardCollection.containers[0].set.items[1].image`
have the same problem and therefore also do not work.

If you trim this extra digit off, you get this URL instead:

https://prod-ripcut-delivery.disney-plus.net/v1/variant/disney/E7BFFF8CD4E7BA85BFB3439CDF90698213E8134E4CC729A9AFA17A2E1FC665D3/scale?format=jpeg&quality=90&scalingAlgorithm=lanczos3&width=500

As you can see, this URL actually works as expected and the page loads.
I haven't closely investigated the other tiles' image sets for the same
typo, but a cursory glance seems to indicate they're unaffected and
appear to work as expected.

As such, the temporary fix is to patch our local `tests/home.json` to
fix the typos, and use that for now. Once the API is fixed, we can
revert this particular change.