Add WidgetId::root() constructor
This change was done so that `State` implementations can contain
`WidgetId` fields that initialize to a well-defined value, instead
of resorting to using `Option::<WidgetId>::None` at construction time,
and setting it later to `Some` in `State::initialize()`. The latter is
annoying because it results in many spurious `unwrap`s down the line,
and this was the best solution I could think of.