Implement up/down scrolling at fixed-height interval
This commit implements up/down scrolling by changing the widget
hierarchy from this:

```
Root
├── Label
│   ├── Tile
│   └── ...
├── Label
│   ├── Tile
│   └── ...
└── ...
```

to this:

```
Root
└── Grid
    ├── Label
    │   ├── Tile
    │   └── ...
    ├── Label
    │   ├── Tile
    │   └── ...
    └── ...
```

`Grid` is a hidden transparent widget to which acts as an anchor point
to all of its child widgets, conveniently allowing the entire widget
hierarchy to be freely translated up/down/left/right without touching
the fullscreen `Root` widget which provides the background color.

The current scrolling algorithm is rudimentary at best: if the user
uses the arrow keys to select a tile whose top-left corner is below
`window height / 2`, we scroll down `ROW_HEIGHT` pixels such that the
selected panel is now at the top of the screen and fully visible.
Similarly, if the user hovers over a tile above `window height / 2`, we
scroll upward `ROW_HEIGHT` pixels instead to achieve the same thing.

This approach technically works, but is still a bit flawed. For one, it
is heavily dependent on the dimensions of the window; if the window is
small or is initialized with an odd resolution, the selected tile may
not be perfectly vertically centered. Another issue is that due to the
fixed height scroll intervals, we can't stop scrolling precisely at the
last row once it comes into view (unless the window height is perfectly
divisible by `ROW_HEIGHT`, of course).

Ideally, we should have a scroll threshold that scales relative to the
window height to keep the selected tile more visible at all times, and
we should have the ability to scroll only `ROW_HEIGHT - delta` once the
final row is fully visible on the screen. These improvements will have
to come in a separate commit.