Implement per-row left/right scrolling
This particular scrolling behavior is much more akin to the interface of
an actual streaming service like Disney+ or Netflix, with the entire
interface scrolling up/down in lockstep, but tiles within the currently
selected row can scroll left/right freely.

Below is a quick-and-dirty diagram of an example menu with two rows, the
top one is scrolled 1 unit to the left, and the other is scrolled 1 unit
to the right.

```text
             -1 | 0 | 1 | 2 | 3 | 4 | 5 | 6
            ----+---+---+---+---+---+---+---+
                |   |
            +---+---+---+---+---+---+
[Row 0]     | 0 | 1 | 2 | 3 | 4 | 5 |
Offset: -1  +---+---+---+---+-+-+---+
                |   |         |
                |   +---+---+-v-+---+---+---+
[Row 1]         |   | 0 | 1 | 2 | 3 | 4 | 5 |
Offset: +1      |   +---+---+---+---+---+---+
                |   |
```

Suppose you wanted to navigate from tile (0, 4) to the (1, 2) tile
immediately below it, as indicated by the arrow above, using only the
numbers in the diagram above.

Computing the row index from [Row 0] to [Row 1] is easy, just 0 + 1 = 1.
Computing the column index is trickier, but the formula is:

```text
cur_col + cur_scroll_offset - dest_scroll_offset
```

So, for example, to move from (0, 4) to the adjacent tile below, one can
compute:

```text
(0 + 1, 4 + (-1) - 1) -----> (1, 2)
```

Therefore, tile (1, 2) in the menu grid is now our newly selected tile.